requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
colorama==0.4.4
dateparser==1.0.0
idna==3.2
multidict==5.1.0
python-binance==1.0.12
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
requests==2.26.0
six==1.16.0
tradingview-ta==3.2.9
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.6
websockets==9.1
yarl==1.6.3
numpy~=1.21.2
prettytable==2.2.1
pandas==1.3.5
pandas-ta==0.3.14b0
ccxt==1.58.23
asyncio==3.4.3
redis
matplotlib
altair
streamlit
tabulate
psutil
pytz
websocket_client
pdoc3